+++ packages/app_intents_codegen/lib/src/models/models.dart
@@ -4,5 +4,6 @@ library;
 export 'entity_info.dart';
 export 'enum_info.dart';
 export 'intent_info.dart';
+export 'snippet_info.dart';
 export 'union_info.dart';
 export 'widget_configuration_info.dart';